# specs/1-obe-system/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: OBE System API
  description: |
    Outcome-Based Education Management System API
    
    This API provides endpoints for:
    - Authentication (JWT, Google OAuth, Email OTP)
    - User management with RBAC
    - Course and outcome management
    - Assessment and result tracking
    - Analytics and reporting
  version: 1.0.0
  contact:
    name: OBE System API Support
    email: support@obesystem.com
servers:
  - url: http://localhost:8000/api
    description: Local development server
  - url: https://api.obesystem.com/api
    description: Production server
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List users (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [Admin, Teacher, Student]
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Users]
      summary: Create user (Admin only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /courses:
    get:
      tags: [Courses]
      summary: List courses (role-based filtering)
      security:
        - BearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of courses
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseListResponse'

    post:
      tags: [Courses]
      summary: Create course (Admin only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCourseRequest'
      responses:
        '201':
          description: Course created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /results:
    post:
      tags: [Results]
      summary: Create result entry (Teacher)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateResultRequest'
      responses:
        '201':
          description: Result created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /analytics/course/{id}:
    get:
      tags: [Analytics]
      summary: Course analytics (Teacher/Admin)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Course analytics data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseAnalyticsResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
  /analytics/program:
    get:
      tags: [Analytics]
      summary: Program-level analytics (Admin)
      security:
        - BearerAuth: []
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Program analytics data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgramAnalyticsResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
  /analytics/student/{id}:
    get:
      tags: [Analytics]
      summary: Student progress analytics (Student/Teacher)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Student progress data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentAnalyticsResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
  /analytics/outcomes/{id}:
    get:
      tags: [Analytics]
      summary: Outcome achievement analytics
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
            enum: [PLO, CLO]
      responses:
        '200':
          description: Outcome achievement data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutcomeAnalyticsResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserResponse:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        role:
          $ref: '#/components/schemas/RoleResponse'
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
    RoleResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
    CreateUserRequest:
      type: object
      required: [email, first_name, last_name, role_id]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        first_name:
          type: string
        last_name:
          type: string
        role_id:
          type: integer
...
    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        total:
          type: integer
        page:
          type: integer
        per_page:
          type: integer
        total_pages:
          type: integer
    CreateCourseRequest:
      type: object
      required: [code, name]
      properties:
        code:
          type: string
        name:
          type: string
        description:
          type: string
        credits:
          type: integer
...
    CourseResponse:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        description:
          type: string
        credits:
          type: integer
        teachers:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        created_at:
          type: string
          format: date-time
    CourseListResponse:
      type: object
      properties:
        courses:
          type: array
          items:
            $ref: '#/components/schemas/CourseResponse'
        total:
          type: integer
        page:
          type: integer
        per_page:
          type: integer
...
    CLOResponse:
      type: object
      properties:
        id:
          type: integer
        course_id:
          type: integer
        code:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time
...
    AssessmentResponse:
      type: object
      properties:
        id:
          type: integer
        course_id:
          type: integer
        name:
          type: string
        type:
          type: string
        description:
          type: string
        weight:
          type: number
        max_score:
          type: number
        due_date:
          type: string
          format: date-time
        clos:
          type: array
          items:
            $ref: '#/components/schemas/CLOResponse'
        created_at:
          type: string
          format: date-time
    CreateResultRequest:
      type: object
      required: [student_id, assessment_id, clo_id, score, max_score]
      properties:
        student_id:
          type: integer
        assessment_id:
          type: integer
        clo_id:
          type: integer
        score:
          type: number
        max_score:
          type: number
...
    ResultResponse:
      type: object
      properties:
        id:
          type: integer
        student:
          $ref: '#/components/schemas/UserResponse'
        assessment:
          $ref: '#/components/schemas/AssessmentResponse'
        clo:
          $ref: '#/components/schemas/CLOResponse'
        score:
          type: number
        max_score:
          type: number
        percentage:
          type: number
        feedback:
          type: array
          items:
            $ref: '#/components/schemas/FeedbackResponse'
        created_at:
          type: string
          format: date-time
    FeedbackResponse:
      type: object
      properties:
        id:
          type: integer
        result_id:
          type: integer
        comment:
          type: string
        created_by:
          $ref: '#/components/schemas/UserResponse'
        created_at:
          type: string
          format: date-time
    CourseAnalyticsResponse:
      type: object
      properties:
        course_id:
          type: integer
        total_students:
          type: integer
        outcome_achievement:
          type: array
          items:
            type: object
            properties:
              clo_id:
                type: integer
              achievement_rate:
                type: number
        average_grade:
          type: number
    ProgramAnalyticsResponse:
      type: object
      properties:
        plo_achievement:
          type: array
          items:
            type: object
            properties:
              plo_id:
                type: integer
              achievement_rate:
                type: number
        total_courses:
          type: integer
        total_students:
          type: integer
    StudentAnalyticsResponse:
      type: object
      properties:
        student_id:
          type: integer
        courses:
          type: array
          items:
            type: object
            properties:
              course_id:
                type: integer
              grade:
                type: number
              outcome_achievement:
                type: array
                items:
                  type: object
    OutcomeAnalyticsResponse:
      type: object
      properties:
        outcome_id:
          type: integer
        type:
          type: string
        achievement_rate:
          type: number
        student_count:
          type: integer
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
  responses:
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...